requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
defusedxml==0.7.1
Django==5.2.8
django-debug-toolbar==4.2.0
gunicorn==23.0.0
idna==3.11
oauthlib==3.3.1
openpyxl==3.1.5
pillow==12.1.1
psycopg[binary]==3.3.3
pycparser==3.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
reportlab==4.4.9
requests==2.32.5
requests-oauthlib==2.0.0
sib-api-v3-sdk==7.6.0
six==1.17.0
social-auth-app-django==5.4.0
social-auth-core==4.8.3
sqlparse==0.5.5
urllib3==2.6.3
